requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
google-generativeai==0.7.2
slack_sdk==3.30.0
notion-client==2.3.0
pydantic==2.8.2
